refactor(Episode): remove Episode and EpisodeStore models; update imports in RootStore and index files; add AuthenticationStore and Project models
