+++ app/models/index.ts
@@ -1,9 +1,7 @@
 export * from "./RootStore"
-export * from "./AuthenticationStore"
-export * from "./Episode"
-export * from "./EpisodeStore"
-export * from "./Project"
-export * from "./ProjectStore"
+export * from "./auth/AuthenticationStore"
+export * from "./project/Project"
+export * from "./project/ProjectStore"
 export * from "./helpers/getRootStore"
 export * from "./helpers/useStores"
 export * from "./helpers/setupRootStore"
